server-model/requirements.txt
mysql-connector-python
scikit-learn
tensorflow==2.18.0
sqlalchemy
pandas
python-dotenv
flask==3.1.0
numpy